requirements.txt
opencv-contrib-python==4.11.0.86
Pillow
numpy
django
